requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
uvicorn[standard]
fastapi==0.103.2
requests==2.31.0
playwright==1.39.0
python-dotenv==1.0.0
pydantic==1.10.9
sqlalchemy==2.0.22
pymysql==1.1.0
python-multipart==0.0.6
pillow==9.5.0
contextvars==2.4
